+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   `{:reply, frame, state}` callback response.
 - Send a close frame with the close code `1011` when there is an unexpected
   error. (Like an `Exception` in the middle of a callback)
+- Add a more specific error when the `:websockex` application hasn't been
+  started yet.
 
 ### Bug Fixes
 - Fix a couple of places where the call stack wasn't being properly tail-call